+++ src/Symfony/Component/Intl/ResourceBundle/Transformer/CompilationContext.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Intl\ResourceBundle\Transformer;
 
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Intl\ResourceBundle\Compiler\BundleCompilerInterface;
+use Symfony\Component\Intl\ResourceBundle\Scanner\LocaleScanner;
 
 /**
  * Default implementation of {@link CompilationContextInterface}.
@@ -32,7 +33,7 @@ class CompilationContext implements CompilationContextInterface
     private $binaryDir;
 
     /**
-     * @var FileSystem
+     * @var Filesystem
      */
     private $filesystem;
 
@@ -46,13 +47,19 @@ class CompilationContext implements CompilationContextInterface
      */
     private $icuVersion;
 
-    public function __construct($sourceDir, $binaryDir, Filesystem $filesystem, BundleCompilerInterface $compiler, $icuVersion)
+    /**
+     * @var LocaleScanner
+     */
+    private $localeScanner;
+
+    public function __construct($sourceDir, $binaryDir, Filesystem $filesystem, BundleCompilerInterface $compiler, $icuVersion, LocaleScanner $localeScanner = null)
     {
         $this->sourceDir = $sourceDir;
         $this->binaryDir = $binaryDir;
         $this->filesystem = $filesystem;
         $this->compiler = $compiler;
         $this->icuVersion = $icuVersion;
+        $this->localeScanner = $localeScanner ?: new LocaleScanner();
     }
 
     /**
@@ -93,5 +100,13 @@ class CompilationContext implements CompilationContextInterface
     public function getIcuVersion()
     {
         return $this->icuVersion;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getLocaleScanner()
+    {
+        return $this->localeScanner;
     }
 }