+++ src/Symfony/Component/Intl/CHANGELOG.md
@@ -8,3 +8,5 @@ CHANGELOG
    whenever an invalid locale is given
  * [BC BREAK] the various Intl methods now throw a `NoSuchEntryException`
    whenever a non-existing language, currency, etc. is accessed
+ * the available locales of each resource bundle are now stored in a generic
+   "misc.res" file in order to improve reading performance